working reach router in place! moved all components into their own files
