+++ app/src/components/About.js
@@ -1,0 +1,14 @@
+import * as React from "react";
+
+import Nav from "./Nav";
+
+const About = () => {
+  return (
+    <div>
+      <Nav />
+      This is About us
+    </div>
+  );
+};
+
+export default About;
