+++ app/src/components/AddTask.js
@@ -1,0 +1,30 @@
+import * as React from "react";
+
+import * as apiClient from "../apiClient";
+
+const AddTask = ({ loadTasks }) => {
+  const [task, setTask] = React.useState("");
+
+  const canAdd = task !== "";
+
+  const onSubmit = async (e) => {
+    e.preventDefault();
+    if (canAdd) {
+      await apiClient.addTask(task);
+      loadTasks();
+      setTask("");
+    }
+  };
+
+  return (
+    <form onSubmit={onSubmit}>
+      <label>
+        New task:{" "}
+        <input onChange={(e) => setTask(e.currentTarget.value)} value={task} />
+      </label>
+      <button disabled={!canAdd}>Add</button>
+    </form>
+  );
+};
+
+export default AddTask;
