+++ app/src/components/TaskList.js
@@ -1,0 +1,11 @@
+import * as React from "react";
+
+const TaskList = ({ tasks }) => (
+  <ul>
+    {tasks.map(({ id, name }) => (
+      <li key={id}>{name}</li>
+    ))}
+  </ul>
+);
+
+export default TaskList;
