+++ package.json
@@ -33,6 +33,7 @@
   },
   "homepage": "https://github.com/gsong/express-react-project-example#readme",
   "dependencies": {
+    "@reach/router": "^1.3.4",
     "npm-run-all": "^4.1.5",
     "rimraf": "^3.0.2"
   }
